requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain-openai==0.1.23
langchain==0.2.16
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
duckduckgo-search==3.9.6